requirements.txt
bleach==6.1.0
blinker==1.7.0
click==8.1.7
Flask==3.0.0
itsdangerous==2.1.2
Janome==0.5.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.24
tomli==2.0.1
typing_extensions==4.9.0
webencodings==0.5.1
Werkzeug==3.0.1
